file cleanup and move dialog classes to new file
